feat(key.go): add support for creating a PemKey from a string key value
The `NewPemKeyFromValue` function is added to the `key.go` file. This function allows creating a new `PemKey` from a string key value. If a passphrase is provided, it will be used to update the passphrase of the `PemKey`. This change improves the flexibility of creating `PemKey` objects by allowing them to be created from a string key value instead of only from a file path.